+++ starsky/starsky.foundation.accountmanagement/Middleware/NoAccountLocalhostMiddleware.cs
@@ -24,10 +24,16 @@ namespace starsky.foundation.accountmanagement.Middleware
 
         internal const string Identifier = "mail@localhost";
 
+        /// <summary>
+        /// Enable: app__NoAccountLocalhost
+        /// </summary>
+        /// <param name="context"></param>
         public async Task Invoke(HttpContext context)
         {
 	        var isHostLocal = IsLocalhost.IsHostLocalHost(context.Connection.LocalIpAddress, context.Connection.RemoteIpAddress);
-	        var isApiCall = context.Request.Path.HasValue && context.Request.Path.Value.StartsWith("/api");
+	        var isApiCall = context.Request.Path.HasValue && (context.Request.Path.Value.StartsWith("/api") || 
+		        context.Request.Path.Value.StartsWith("/realtime"));
+	        
 	        var isFromLogoutCall = context.Request.QueryString.HasValue && 
 	                               context.Request.QueryString.Value.Contains("fromLogout");
 	        
@@ -35,16 +41,17 @@ namespace starsky.foundation.accountmanagement.Middleware
 	        {
 		        var userManager = (IUserManager) context.RequestServices.GetService(typeof(IUserManager));
 
-		        var users = userManager.GetUser("email", Identifier);
-		        if ( users == null )
+		        var user = userManager.GetUser("email", Identifier);
+		        if ( user == null )
 		        {
 			        var newPassword = Convert.ToBase64String(
 					        Pbkdf2Hasher.GenerateRandomSalt());
 			        await userManager.SignUpAsync(string.Empty, 
 				        "email", Identifier, newPassword+newPassword);
+			        user = userManager.GetUser("email", Identifier);
 		        }
 		        
-		        await userManager.SignIn(context, users, true);
+		        await userManager.SignIn(context, user, true);
 	        }
             await _next.Invoke(context);
         }