+++ starsky/starsky.foundation.database/Migrations/20211228192706_ImageStabilisation.cs
@@ -1,0 +1,23 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+namespace starsky.foundation.database.Migrations
+{
+    public partial class ImageStabilisation : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<int>(
+                name: "ImageStabilisation",
+                table: "FileIndex",
+                nullable: false,
+                defaultValue: 0);
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "ImageStabilisation",
+                table: "FileIndex");
+        }
+    }
+}