+++ starsky/starsky.foundation.database/Models/ImageStabilisationType.cs
@@ -1,0 +1,9 @@
+namespace starsky.foundation.database.Models
+{
+	public enum ImageStabilisationType
+	{
+		Unknown = 0,
+		On = 1,
+		Off = 2
+	}
+}