+++ starsky/starskytest/Services/ReadMeta_ExifReadTest.cs
@@ -7,8 +7,10 @@ using MetadataExtractor.Formats.Iptc;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using starsky.foundation.database.Models;
 using starsky.foundation.platform.Helpers;
+using starsky.foundation.readmeta.Helpers;
 using starsky.foundation.readmeta.ReadMetaHelpers;
 using starsky.foundation.readmeta.Services;
+using starsky.foundation.storage.Storage;
 using starskycore.Attributes;
 using starskytest.FakeCreateAn;
 using starskytest.FakeMocks;
@@ -133,6 +135,41 @@ namespace starskytest.Services
 			Assert.AreEqual( "Sony", item.Make);
 			Assert.AreEqual( "SLT-A58", item.Model);
 			Assert.AreEqual( "24-105mm F3.5-4.5", item.LensModel);
+			Assert.AreEqual( ImageStabilisationType.Unknown, item.ImageStabilisation);
+
+		}
+		
+		[TestMethod]
+		public void ImageStabilisationOn()
+		{
+			var newImage = CreateAnImageA6600.Bytes;
+			var fakeStorage = new FakeIStorage(new List<string>{"/"},
+				new List<string>{"/test.jpg"},new List<byte[]>{newImage});
+			 
+			var item = new ReadMetaExif(fakeStorage).ReadExifFromFile("/test.jpg");
+			Assert.AreEqual( ImageStabilisationType.On, item.ImageStabilisation);
+		}
+		
+		[TestMethod]
+		public void ImageStabilisationOff()
+		{
+			var newImage = CreateAnImageA58Tamron.Bytes;
+			var fakeStorage = new FakeIStorage(new List<string>{"/"},
+				new List<string>{"/test.jpg"},new List<byte[]>{newImage});
+			 
+			var item = new ReadMetaExif(fakeStorage).ReadExifFromFile("/test.jpg");
+			Assert.AreEqual( ImageStabilisationType.Off, item.ImageStabilisation);
+		}
+				
+		[TestMethod]
+		public void LensModelTamRon()
+		{
+			var newImage = CreateAnImageA58Tamron.Bytes;
+			var fakeStorage = new FakeIStorage(new List<string>{"/"},
+				new List<string>{"/test.jpg"},new List<byte[]>{newImage});
+			 
+			var item = new ReadMetaExif(fakeStorage).ReadExifFromFile("/test.jpg");
+			Assert.AreEqual( "Tamron or Sigma Lens", item.LensModel);
 		}
 
 		[TestMethod]